move dataclasses to energy_demand
